Make QifParser recognize investment accounts. Note: doesn't actually work.
QifParser, and hence fixofx, will no longer throw an exception when parsing QIF investment statements. A valid OFX document _of the wrong type_ will be returned instead. The OFX document will make it appear that the found transactions are bank account transactions instead of investment transactions.

To make full QIF investment parsing work, we'd have to add investment support throughout lib/ofx and lib/ofxtools. Both of those are doable but I haven't done the work on that. If someone else wants to take it on, go for it.